Update pipe.js
This is to increase the timeout from 20 to 60 seconds.  
We have larger dashboards with history and really prefer to keep our current setup.  Chrome only uses a single cpu to load the dashboard; Firefox is better with multiple vCPU but not all my users are allowed to install Firefox.  Users with lower clock speeds have issues loading dashboards that have a lot of history.  Users with slow VPN also receive this timeout.

JENKINS-27354
JENKINS-64253
https://stackoverflow.com/questions/25481734/jenkins-delivery-pipeline-error-error-communicating-to-server